requirements.txt
google-genai
pyttsx3
python-dotenv
apscheduler
